Allow explicit value-blocking in `xcss` prop.
We don't actually need the generic value for this to be valuable, this should block it so an `@compiled/react.cssMap` with conflicting values is disallowed, but also can allow prop-level restrictions (if we want).

This is by moving `CompiledPropertyDeclarationReference` to not be optional as that object seemingly messed something up as `| CompiledPropertyDeclarationReference` was always "truthy" (if that makes sense).

However, this removes the functionality for inline `xcss` and requires a `cssMap` or future `css`(?) to exist..which is probably fine.

---

This has one remaining scenario not erroring, but it does error somewhere in the stack, and this is because a cssMap-level type error is typed differently than a non-erroring cssMap?