Allow `ax()` to handle non-string classnames when collected.
The example being passing `<Component xcss={cx(…)} />` or `<Component xcss={[…]} />` results in `ax` capturing things like this:
```tsx
[
  '_1bah1yb4',
  undefined,
  ['_16jlkb7n', undefined],
  AtomicGroups {
    values: Map(2) { '_1bsb' => '_1bsb1osq', '_16jl' => '_16jlkb7n' }
  }
]
```

I don't know if it's the best way to handle this, but internally we've just been invoking `toString()` via a wrapper, and we're trying to get to native Compiled.  Makes sense (and works) to me!